+++ src/jsp/com/cubrid/jsp/jdbc/CUBRIDServerSideConnection.java
@@ -33,7 +33,7 @@ package com.cubrid.jsp.jdbc;
 
 import com.cubrid.jsp.ExecuteThread;
 import com.cubrid.jsp.data.DBParameterInfo;
-import com.cubrid.jsp.io.SUConnection;
+import com.cubrid.jsp.impl.SUConnection;
 import cubrid.jdbc.jci.CUBRIDIsolationLevel;
 import java.io.IOException;
 import java.sql.Array;
