+++ src/jsp/com/cubrid/jsp/jdbc/CUBRIDServerSideStatement.java
@@ -31,7 +31,7 @@
 
 package com.cubrid.jsp.jdbc;
 
-import com.cubrid.jsp.io.SUStatement;
+import com.cubrid.jsp.impl.SUStatement;
 import cubrid.jdbc.driver.CUBRIDJDBCErrorCode;
 import cubrid.jdbc.jci.CUBRIDCommandType;
 import java.io.IOException;
